chase: introduce flags that verify that chased inode is regular file or dir
This also implies the new CHASE_MUST_BE_DIRECTORY flag in case the
specified path ends in a slash. This makes the rules stricter, it means
we'll be closer to how this is handled in kernel: if a path ends in a
slash it can never refer to a non-directory.

(cherry picked from commit 90b9f7a07e6f57825f416f6ce2db0a9f2086754b)

Resolves: RHEL-75774